IMAP: Fix SPECIAL-USE mailbox calculation.
The original algorithm to compute the SPECIAL-USE mailboxes
was incorrect when multiple SPECIAL-USE mailboxes are
present (e.g. multiple mailboxes in different namespaces),
resulting in operations like "Delete", "Junk", etc. moving
messages to the wrong folder.

The logic has been amended to be more careful to ensure
the right mailboxes are calculated.