editor: Restore original cursor position after updates.
Fix an occasional annoyance where the cursor would get
moved as a byproduct of updating another window, while
the cursor is active, leaving it on the other window
instead of in the original position. To fix this, simply
mark the original window as needing an update so that the
cursor is restored to the original position.