evergreen: Allow creating and deleting folders.
Allow using IMAP CREATE, DELETE, and RENAME to create
new mailboxes or delete existing ones. Arbitrary
mailbox renaming is not yet supported.

This implementation is not super optimal at the moment,
as after these operations, the LIST command is effectively
repeated in order to repopulate the new mailboxes array.
This could be optimized significantly in the future,
and a stub has been left for this in imap.c.