imap.c: Fix mailbox sorting algorithm.
The original method of sorting mailbox names did not always sort
folders in the canonical order, and in particular would lead to
subfolders of special folders not sorting immediately after their
parents, leaving them depicted under the wrong parent when this
ordering was used graphically. Redo the logic to always sort folders
in the expected order.