imap.c: Optimize CREATE/DELETE/RENAME operations.
Instead of reissuing a LIST command to rebuild the ordering of the new
set of mailboxes, build the new mailboxes array from the old array,
inserting/deleting the created/deleted/renamed mailbox in the right
spot. This reduces the delay following these operations in the program
from potentially as long as several seconds to almost instant.