imap.c: Add support for RFC 5465 NOTIFY.
Add support for NOTIFY, so that we can receive updates about
new mail in folders besides the currently selected one. We
already handled STATUS in IDLE responses, so we just needed
to add the code to enable the NOTIFY extension appropriately.

As part of this change, we also treat INBOX folders case-insensitively,
as in https://github.com/InterLinked1/wssmail/commit/e4548597fc67e6f322f6fe09afc17b456e3f2169.

This change (and the referenced one) are needed to account for
scenarios such as when Other Users.example.Inbox appear in the
LIST response and navigation, but STATUS updates use a different
case, e.g. Other Users.example.INBOX.