evergreen.c: Add support for renaming folders.
Add general support for renaming folders. Previously, the only
renames supported were when folders were moved to the trash
as a subfolder.

The user is required to add or remove the hierarchy delimiter
manually if moving the folder (as opposed to renaming it where
it is), so the user is reminded of this prior to being prompted
for the new name.

The key F12 has been assigned to "rename folder", since F2 cannot
be detected reliably across terminals.

The confirmation status for create/delete/rename operations has
also been improved.