evergreen.c: Improve display of help options and html2text warning.
Many help options are paired with a complementary option (e.g. previous,
next); however, some options near the bottom were not properly paired up.
Add an empty filler field to better align up these options to their
corresponding paired actions.

Additionally, a warning is triggered if html2text is not present, but
this was not visible since the status bar was immediately overwritten
with message metadata (we also weren't calling doupdate). This has been
adjusted so the error is briefly visible to the user.